quilt: Add option for TLS-encrypted GRPC communication
Before, we used insecure HTTP for our GRPC servers, which meant that
malicious parties could read our traffic, or even masquerade as the
controller and change specs at will.

This patch adds support for TLS mutual authentication for communicating
with the minion and daemon. If TLS is enabled, the minion servers boot
with a public, private key pair signed by a given certificate authority.
To connect, clients must have a key pair signed by the same CA. Clients
will also refuse to connect to servers that lack a certificate signed by
the CA.

Note that TLS must be explicitly enabled, so this patch will not affect
the current workflow.