daemon: Automatically grant an admin key access to machines
This commit adds an `admin-ssh-private-key` flag to the daemon so that
users can specify a private key that machines should always allow access
from.  This is necessary for TLS encryption so that the cluster can
install TLS certificates on cloud machines.

This flag is also useful for users that don't want to specify SSH keys
in the blueprint -- they can leave the `sshKeys` field of machines
blank, and simply access the cluster using the `admin-ssh-private-key`.