counters: Implement the new counters package
A simple event counter can be an incredibly helpful debugging tool.
They're cheap enough to always be running, even in production.
They're unobtrusive to the average user.  Yet, they can convey a bunch
of information about the behavior of the system that would otherwise
be completely opaque.

This patch implements a new package, counters for Quilt.  It also
implements a new command, `quilt counters` which can query counters
from the daemon (querying from the minions is not yet implemented).