connection: Pull out logic to obtain GRPC clients and servers
This patch pulls out the code for obtaining GRPC clients and servers into
the `connection` package.

Although creating clients and servers is trivial right now, the code
will become more complicated once we have different authentication
methods. At that point, it will be helpful to only have to modify the
grpc code in the connection package.