stitch: Let blueprints write to stdout
Before, we gathered the deployment object by printing the object to
stdout. However, if the blueprint used `console.log` to print anything
to stdout, the output would get bundled with the deployment object,
resulting in a parsing error.

This commit solves this by communicating the deployment object by
writing it to a random file on the filesystem. This way, we impose no
restrictions on stdout.