js: Add ESLint
This commit adds functionality to lint our JavaScript code
using ESLint and Google's style guide. For now, the errors
about missing JSDoc are downgraded to warnings in bindings.js,
which is currently missing lots of JSDoc. Also, one of the
test files calls code from another repository that doesn't
comply with the style guide, so until we fix that code, that
issue is marked as a warning rather than an error.