cluster: Refactor updateCloud to not rely on an enum
Before, `updateCloud` took an argument that determined what provider
actions should be done, and what log messages should be outputted. This
patch simplifies `updateCloud` to take the provider action function as an
argument directly. This removes the need for the `action` enum, allowing
us to remove a lot of ugly switching code.