ps: Only query containers if machines have been connected to
Before, `quilt ps` displayed a scary warning when the cluster was
not fully booted.

This patch makes it so `quilt ps` only attempts to display container
information if any machine has booted enough for the foreman to connect
to it. This is a valid heuristic because if the foreman hasn't connected
to the machine, there's no way for it to receive the deployment object
to specify what containers should be booted.