build: Fix format checking
The Makefile used to have 3 format-related targets: format, lint,
and format-check.  These had become intertwined over time; e.g.,
the lint command would fix errors in the code (and depended on
format), meaning that if Jenkins called lint, it wouldn't fail
for some format errors (instead, it would fix the errors in the
version of the code running on Jenkins). This commit eliminates
the format command, and updates lint to only include functionality
to check the format of code.  Now, the format make target is the
only target that will update code to comply with the correct format.