ctl: Rename `ps` to `show` while keeping `ps` as an alias
Most people don't know what `ps` is, making the `quilt ps` command
seem jargony and inaccessible.  `show` is much more descriptive and
should be the default.  That said, `ps` is used in other tools, and
some users may expect it to work.  So this patch allows either to be
used.