openflow: Embed Container in container
The openflow.container is just an internal representation of
openflow.Container with some additional context.  In future patches,
as more stuff is added to openflow.Container, adding it to both
structs will become a burden.  Thus, this patch embeds
openflow.Container in openflow.container so the internal
representation has easy access to the external representation's fields.