ssh: Properly log failures to load default keys
Before, we logged failures to load default keys at Debug because we
expected there to be failures from keys missing. However, these errors
were sometimes related to parsing keys, which should be shown to the
user.

This patch modifies the error handling to log at Debug when a file isn't
found, and log at Warn otherwise.