quiltctl: Only list options once in usage strings
Before, we would list command options both in the `Usage: quilt ...` string,
and again at the bottom of the usage string, using the FlagSets print function.
This both made the usage strings harder to read and unnecessarily long,
and led to discrepancies between the actual flags and those listed at the top of
the usage string.

We now just say [OPTIONS] at the top, and then list the registered flags at the
bottom.