registry: Don't use image cache when building images
Before, if an image was rebuilt with the same Dockerfile, we would use
the cached version of the image and not actually rebuild it. This was
confusing for images that interacted with the network such as the
website because it wouldn't pull in the latest website files.

For now, this commit changes the building logic to never use the cache.
Note that we do not rely on the caching logic when there are multiple
containers with the same Dockerfile, so this change will not slow down
the system in the case where multiple copies of a container with a
custom Dockerfile are deployed.