make: Don't run `govendor` when creating coverage targets
Because the `go-coverage` target depends the PACKAGES variable, make
was calling govendor whether or not that target was eventually
evaluated.  The govendor tool may not be installed when a user first
runs `make go-get`, so this would cause spurious warnings.  This patch
pulls the calculation of targets into the body of `go-coverage`,
causing it only to be evaluated when needed.

Reported-By: Diman Todorov <diman.todorov@gmail.com>