js: Ensure container hostnames are unique by appending IDs
Before, libraries that deployed containers with hostnames had to
internally ensure that they never deployed two containers with the same
hostname. This situation is common when users expect calling the library
twice will return two separate deployments, rather than two identical
deployments with the same hostnames.

As a side-effect, this change also ensures that a container and label
will not be assigned the same hostname. Before, it was possible to
create a Service and Container with the same hostname.