js: Remove invariant support
This commit removes support for invariants from the Quilt Javascript
library. Specifically, it removes the `enough` type, and the
`Deployment.assert`, `Service.annotate`, `Service.canReach`,
`Service.canReachACL`, `Service.between`, and `Service.neighborOf`
methods.

The invariants feature is not used by our current specs, is not well
tested, and requires a non-trivial amount of overhead to maintain given
the changes we're making to the Quilt library.

This commit also moves the remaining graph code into the `inspect`
package because that's now the only caller..