js: Remove LabelRule
This commit removes support for the LabelRule placement rule. Placement
was poorly designed, so we're removing it for now until we have a better
proposal. Furthermore, no blueprints relied on `LabelRule` placements,
so removing it from the library will not break any existing blueprints.

The scheduler still retains a concept of placement rules based on
service groups so that containers can be placed on different machines
when they listen for public connections on the same port.

Machine placement rules were not touched.