js: Rename machine rule
This commit simplifies the API for specifying container-machine
placement rules. It renames the `place` method to be `placeOn`, and
removes the `exclusive` boolean argument. We can do this because we
currently only support placing containers on machine types, so the
previous flexibility of specifying "exclusivity" and placement types is
not needed.

Specifically, `Service.place(false, new MachineRule({floatingIp}))` is
now `Service.placeOn({floatingIp})`.