quiltctl: Add init command
This commit adds a `quilt init` command that will prompt the user for
details about their infrastructure and as a result create a blueprint that
describes an infrastructure with the given specifications.

The infrastructure will be stored in ~/.quilt/infra/infra_name.js where
infra_name is the user given name. The file exports a function that when called
returns a deployment object.

A future commit will add more extensive documentation on usage and the prompts,
as well as update the existing getting started guide to use `quilt init`.