digitalocean: Pin the base Ubuntu image
Before, we referenced the image by a friendly name that always pointed
to the most recent Ubuntu image. This was problematic because we had no
control over when the Ubuntu image would update and potentially break
Quilt. Furthermore, the kernel version would frequently update, requiring
us to update our OVS container to package the new version of the kernel
module.

We now reference a 16.04.1 image directly by ID.

At the time of this commit, the friendly name image pointed to Ubuntu
version `16.04.3 LTS` with the kernel `4.4.0-87-generic`, and was
updated just a week ago.

The pinned image in comparison is version `16.04.1 LTS` with
kernel `4.4.0-62-generic`, and was last updated
about a half year ago.

This change will hopefully fix the failing integration tests on
DigitalOcean.