js: Require all containers to have a hostname
Once we express network connections in terms of hostnames, a container
will have to have a hostname in order to be connected. This is a
reasonable restriction because if a container did not have a hostname,
applications would not be able to access it anyways.

The hostname must be provided as the first argument in the Container
constructor:
```
new Container('hostname', 'image', {
  command: ['my', 'args'],
});
```