scheduler: Define placement in terms of containers
This commit modifies the scheduler to define placement rules in terms of
containers, rather than labels. It also modifies the Javascript API to
be in sync with this change.

This change is natural given the upcoming patches to make containers the
unit of deployment. Otherwise, in order to define a placement rule on a
container, users would have to wrap the container within a Service, and
apply the rule to that Service.