network: Define connections based on hostname
Before, network connections were described in terms of label names.
This, however, will not work once containers are deployable directly
because they won't necessarily be wrapped within a label. This commit
modifies the deployment engine to accept connections described in terms
of hostnames. The Javascript API was also modified so that containers
must be explicitly connected to each other, rather than through services.

Note that because containers are no longer grouped using labels, we can
no longer make use of OVN's address sets. Before, a connection between
two services only required one ACL, but it now requires O(n^2) ACLs
because an ACL must be made for each pair of containers.

To facilitate connecting groups of containers, this commit also makes
two changes to the Javascript API. It modifies `allowFrom` to accept a
list of containers, and it adds the `allow` function. `allow` can take a
list of Containers as the connection source so that users don't have to
loop over all sets of containers themselves.