network: Load balance in terms of hostnames
This commit modifies the load balancer to be described in terms of
hostnames rather than container IDs. This is good because it decreases
our reliance on stitch IDs, is more consistent with the Connection
object, and theoretically allows load balancers to be nested.

This change also allows us to simplify the database representation of
load balancers. It removes the `Labels` column of the `Container` row,
and the `ContainerIPs` column of the Label row. The code path for
inserting rows into the database was very confusing before:  we would
insert containers in `minion/engine`, allocate IPs for those containers
in `minion/network/ip`, and then insert label rows according to the
labels of those containers. We now insert the label rows in
`minion/engine`, and the load balancer code resolves the hostnames
itself.