js: Remove Service.children support
`Service.children` was originally intended to provide a way for
containers to communicate directly with containers within a Service.
This was designed before containers had individual hostnames, and before
we had decided that a Service (previously known as "Label") should just
be used for load balancing.

`Service.children` was a poor decision for multiple reasons:
- It introduced the requirement that the association between the
`Service.children` hostnames and containers remained consistent.
- There was no way to tell from the Javascript the association between
the hostnames and containers -- the association was determined at
runtime.
- Services should just be used for describing a load-balanced set of
containers -- they should not be used for describing connectivity
between the load-balanced containers.