js: Don't use clone() when generating IDs
Before, we assigned Machines and Containers IDs by calling their clone
function so that we would have a clean reference to mutate. This was
necessary because we did not want to include the `refID` in the hash.

This is not ideal for several reasons:
- The `clone` method is not guaranteed to return an instance with the
exact same attributes. For example, an upcoming patch will ensure that
`clone`d containers have a unique hostname.
- There might be private fields that would be copied by `clone`, but
should not be considered as part of the hash. This will be true in an
upcoming patch that stores connection and placement information in the
Container object.

This patch tweaks the ID generation algorithm to call a `hash()` method
that `Container` and `Machine` define.