js: Make container hostname immutable
This commit removes the `Container.setHostname` function. This way, users do not
have to worry about a call to `Container.setHostname` invalidating an
earlier call to `Container.getHostname`. Furthermore, this theoretically
allows the connection object to be fully created when `allowFrom` is
called, rather than after all code has finished running because there is
no chance that the connected hostnames could change after the call.