js: Vet deployment on generated object
Rather than vet before converting the Quilt deployment into JSON, this
commit makes it so the vet takes place after the deployment has been
converted to JSON. This will make it possible to catch bugs in the
conversion to the deployment object, and generally simplifies the
vetting logic as it doesn't need to make assumptions about how the
deployment will be converted into JSON.

This also fixes a bug where containers that are in multiple services
would trigger the vet failure of repeated hostnames.

Note that this commit did not convert the check for floating IPs applied
to a service with multiple containers because this restriction will soon
be removed when placement rules apply directly to containers.