show: Display container hostname rather than label
Soon, containers won't have to be deployed within a Service. It will
then be common for them to not have any labels. The hostname, in
contrast, is required, and contains a name provided by the user, making
it a good way to identify containers.

We remove this reference to labels now so that the labels field can be
entirely removed from db.Container in the next commit, which refactors
how we manage labels.