util: Remove the JavaScript util library
The JavaScript util library was extremely thin.  The functions it
exports could easily be embedded in their callers.  Furthermore, it
was so far from the code that needed it, that packaging up the init
script in `@quilt/install` was proving more complex than necessary.
To simplify things, this patch ditches it.