jslint: Make automated changes to airbnb style
Quite a few of the differences between Airbnb and google style can be
made with `eslint --fix` option.  This patch goes and makes those
changes, but (for the sake of incrementalism) leaves us in the Google
style for now.  There are several rules where Google and Airbnb style
directly conflict.  In those cases, this patch simply disables those
checks until they're re-enabled in a future patch.