network: Log unknown hostname error at Debug
Before, we logged errors about connections in terms of unknown hostnames
at Warning. However, there is sometimes a race condition where the
connections are synced to workers before hostnames mappings. In this
case, the error is expected, and will go away once the mappings are
synced. Furthermore, the Javascript bindings vets that there are no
connections in terms of undeployed hostnames.