js: Add directory for js examples
The intention is to move some of the Quilt examples from being
in their own repositories to being in the new examples directory,
so that they're easier to maintain and keep up to date with Quilt.
This directory should not be used for blueprints that are useful
for others to import and re-use; it should only be used for
illustrative examples.