docs: Move 'How Quilt Works' to later in the docs
In an effort to separate the information on how to use quilt vs. how Quilt
works, we separate the "How Quilt Works" section from the "Getting Started"
guide. An additional benefit is that this more technical, and somewhat more
intimidating section isn't the first thing users see when they scroll through
the docs.