docs: Move info on minion development
Not many people will need this information, so this commit moves
it to the end of the developer documentation.  This commit also
moves it to be a higher level heading, so that it's included
in the Table of Contents (to make it easier for someone to find
it when necessary).