inspect: Refactored file layout
This commit removes cli/command/inspect.go, and folds that functionality
and cli/command/inspect/main.go into a new file.  As part of this
change, this commit refactors the code to be more consistent with how
other commands are defined: the parsing logic is now located in
the command's Parse function, rather than the Run function.