init: Add help strings to `quilt init`
Some of the `quilt init` questions now provide a help string when the user
types '?'. The help strings function as short documentation supplements
for the questions that are the least intuitive.
It is currently only possible to add help strings to questions that ask for user
input (and e.g. not questions that list a number of options).