js: Force namespaces to be lowercase
GCE doesn't accept capitalized namespaces. Before this commit, if a user tried
to use a namespace with a capital latter on GCE, the daemon would log an error
while booting machines. This could be hard to debug, so this commit makes it so
'quilt run' fails when the blueprint specifies a namespace with any capitalized
letters.