join: Fix algorithm to strictly prioritize better matches
Previously, join used an overly greedy algorithm that could, depending
on the order of the input slice, create an 'okay' pair for an element
even if that element could also be involved in a 'perfect' pair. The
new algorithm does not necessarily return in the same order as the
previous algorithm.