tls: Don't allow unencrypted GRPC connections
Before, if no `tls-dir` was supplied, Quilt would still connect to the
cluster using an unencrypted channel. This insecure-by-default behavior
made it too easy for users to run insecurely. Furthermore, the secret
management design will make use of the TLS credentials, so if we allowed
insecure deployments, secrets would not work for those deployments. This
commit makes it so that neither connections to the cluster, nor
connections to the daemon happen without TLS encryption.

To decrease the chances of accidentally booting the cluster with one set
of credentials, but trying to connect with another set, this commit also
removes the `tls-dir` flag. Instead, TLS credentials are always read
from `~/.quilt/tls`.