daemon: Save autogenerated SSH key to disk
Before, we autogenerated a SSH key for installing TLS certificates, but
only kept it in memory. This meant that if the daemon was restarted
before it had a chance to install the TLS certificates on the machines,
it would be unable to access the cluster because it would no longer have
access to an SSH key that allowed access.

This commit also removes the `admin-ssh-private-key` flag as it was a
short-term solution for allowing users to install certificates with a
SSH key they know has access. However, now that we're saving the SSH key
to disk, the daemon can simply always use the generated key.