google: Properly block after modifying access configs
This commit fixes a bug where assigning a Google machine a floating IP
would fail. This was because we would issuing concurrent requests to
remove the old access config for the ephemeral IP, and to add the new
IP. This would sometimes cause there to be two access configs --
something that neither Google nor our code handles gracefully.

This commit also blocks after adding a new access config so that if a
list happens immediately after, we won't be in a state where there are
no access configs attached to a machine.