blueprint: Remove most uses of infra.deploy()
This commit removes all uses of deployment.deploy(x) where x is a
Container or an application that implements a deploy() method, and
changes all of those uses to instead call x.deploy(deployment).
The only remaining uses of deployment.deploy() are when the
parameter is a list of machines (these uses will be removed in a
future commit).