ci: Fix pinned version for Django
The pinned Git hash was for the commit as it was before it was merged.
The hash changed once the PR was merged because Kevin got added as the
Committer (and Kay remained the author). Referencing a commit from a PR
seems to be supported in NPM v4.2.0 but not v5.3.0.

This commit changes the pinned hash to be the commit in the master
branch -- this is more clear, and works in both NPM v4.2.0 and v5.3.0.