eslint: Check for camelCase filenames
The convention in JavaScript is to use CamelCase for filenames;
this commit updates our eslint configuration to error when non-
camelCase filenames are used (e.g., "blueprint_test.js" would cause
the linter to error).

The default match-regex configuration checks for camelCase,
but it's too strict (e.g., it doesn't allow floatingIP.js),
so this commit uses a custom regex.