blueprint: Deprecate createDeployment()
This commit deprecates createDeployment() in favor of directly
using the Deployment() constructor.  As part of that change, this
commit moves the functionality in createDeployment() that
assigns the Deployment to the global Deployment variable into the
Deployment constructor, and marks createDeployment() as deprecated.
This commit also updates all of the JavaScript in this repository
to use new Deployment() instead of createDeployment(). We're
eliminating createDeployment() because it's more intuitive to use
the constructor directly.