cloud: Sync TLS certificate mappings to minions
This commit updates the foreman to send the private IP to TLS
certificate associations to the minions in the cluster. The upcoming
Vault implementation will make use of this information to create access
policies.

Note that this commit removes a unit test that ensures we skip machines
that have already been setup because it's hard to test that
functionality given the new signature of syncCredentialsOnce, and an
upcoming commit will add a more complete version of the test.