tls: Don't overwrite existing cert
Before, we would overwrite existing minion TLS credentials if the daemon
restarted. This was because the code kept track of what machines had
been initialized using an in-memory map. We now check the remote
filesystem directly before installing credentials.

This avoids a variety of subtle bugs where the daemon's view of the
credentials for each machine, and the actual credentials used by the
minion, become inconsistent. For example, if we used the existing
logic for the upcoming secret implementation, minions might be unable to
retrieve secrets out of Vault after the daemon restarts. This is because
the daemon would install new certificates onto the minion and configure
Vault to only allow access from the new certificate, but the minion
would still use the old certificate, and thus be unable to authenticate.