google: Fix unassigning floating IPs
Before this commit, when floating IPs were unassigned from a machine, we
still tagged the IP as a floating IP rather than an IP assigned by
Google. This meant that on the next join the cloud code would attempt to
remove it. This resulted in IPs being continuously added and then
deleted because the removal of the dynamic IP would go through the same
buggy floating IP code path.

This commit fixes this bug by properly tagging IPs generated by Google.