kelda: Implement secure Secret storage
This commit adds support for Secrets -- a Kelda-blessed method for
securely using sensitive values. Before, secrets (such as OAuth tokens)
were stored in the blueprint source. It was very easy to accidentally
commit these secrets, and they had no special protection once deployed
to the cluster. Secrets are now a first-level concept of Kelda. They
are encrypted at rest and in transit, and access to them is limited
based on least privilege.

Secrets are referenced in blueprints by name -- their values must be
defined separately from the blueprint using the `secret` subcommand.

Secrets can be used as values in the env and filepathToContent fields of
Containers:
```
new kelda.Container(hostname, image, {
  env: {
    notASecret: 'foo',
    verySecret: new kelda.Secret('mySecret'),
  },
})
```

The value of 'mySecret' is defined using the `kelda secret` command:
```
$ kelda secret mySecret 'my secret value'
```