scheduler: Stop containers before booting new ones
Before this commit, if the join returned that a container with the same
IP needed to be restarted (i.e. there was a container to be killed and
a container to be booted with the same IP), the boot request would fail
because Docker doesn't allow two running containers to have the same IP.

While this currently doesn't happen because changes to container
attributes result in new IP addresses, the upcoming secret
implementation will result in containers restarting if their secret
value in Vault changes. When this happens, a new IP is not assigned,
triggering this bug.