Makefile: Fail when integration test fails to compile
The previous Makefile did not properly exit when an integration test
failed to compile because it used a for loop to iterate over the
targets, which doesn't properly propagate build errors to `make`. This
commit rewrites the Makefile so that there is a target responsible for
building a single test. This way, `make` can tell that the compilation
failed.

This Makefile also only attempts to compile tests for suites that
actually contain a _test.go file, which avoids an error that would get
logged when it tried to compile suites with no tests, such as
60-duplicate-cluster-setup.