bindings: Remove CPU and RAM from machine attributes
Before this commit, CPU and RAM are machine attributes. This has
two problems.

1. More information has to be passed between JS and Go.
2. A user might modify a machine's CPU and RAM after creating it.

This commit fixes it, by removing CPU and RAM from machine. The
constructor takes in CPU and RAM as arguments and passed them to
chooseSize. Instead of CPU and RAM, JS only needs to pass Size to
Go.