blueprints-tester: Test against the local bindings.js
Before, Travis would test the blueprints against the version of the
bindings on NPM. This would cause the build to fail when the tested
blueprints were changed to be compatible with the not-yet-merged version
of bindings.js. This change will also cause Travis to properly error
when we forget to update our blueprints in response to an API change.