ci: Wait for the credential files rather than just the directory
Commit db2c21b2 (ci: Wait for the daemon to generate TLS credentials)
attempted to avoid a race between the daemon generating the credentials,
and the tester running commands that required those credentials.
However, the presence of simply the TLS directory is not sufficient for
the commands to work -- the actual credential files are needed. The
commands were still failing because the daemon creates the TLS directory
before actually writing the credentials, so the tester would start
running kelda commands while the daemon was writing the credentials.